feat: implement user authentication middleware and payload handling
- Added AuthMiddleware to handle JWT token verification and payload extraction.
- Introduced TokenPayloadParam decorator for easy access to token payload in controllers.
- Updated UsersController and PostsController to utilize the new authentication guard and payload extraction.
- Modified UsersService and PostsService to accept payload data for user and post operations.
- Created PayloadDto to define the structure of the token payload.
- Enhanced error handling in AuthTokenGuard and middleware for better security.

+++ api/src/users/users.service.ts
@@ -5,6 +5,7 @@ import { createId } from '@paralleldrive/cuid2';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { AuthService } from 'src/auth/auth.service';
 import { HashingServiceProtocol } from 'src/auth/hash/hashing.service';
+import { PayloadDto } from 'src/auth/dto/payload.dto';
 
 @Injectable()
 export class UsersService {
@@ -58,7 +59,7 @@ export class UsersService {
     }
   }
 
-  async updateOne(id: string, body: UpdateUserDto) {
+  async updateOne(id: string, body: UpdateUserDto, payloadToken: PayloadDto) {
     try {
       const findUser = await this.prisma.users.findFirst({
         where:{
@@ -67,6 +68,8 @@ export class UsersService {
       })
   
       if(!findUser) throw new HttpException("User not found", HttpStatus.NOT_FOUND)
+
+      if(findUser.id !== payloadToken.sub) throw new HttpException("Access denied", HttpStatus.NOT_FOUND)
 
       const dataUser: {name?: string, passwordHash?: string} = {
         name: body?.name ? body.name : findUser.name
@@ -99,7 +102,7 @@ export class UsersService {
     }
   }
 
-  async deleteOne(id: string) {
+  async deleteOne(id: string, payloadToken: PayloadDto) {
     try {
       const findUser = await this.prisma.users.findFirst({
         where:{
@@ -108,6 +111,8 @@ export class UsersService {
       })
   
       if(!findUser) throw new HttpException("User not found", HttpStatus.NOT_FOUND)
+
+      if(findUser.id !== payloadToken.sub) throw new HttpException("Access denied", HttpStatus.NOT_FOUND)
       
       await this.prisma.users.delete({
         where: {
